added cpu util

+++ src/processor.cpp
@@ -17,6 +17,6 @@ float Processor::Utilization()
     std::cerr<<t_d<<" "<<i_d<<std::endl;
     prev_idle=idle;
     prev_total=tt;
-    std::cerr<<(float)(1.0-(t_d/i_d))<<std::endl;
-    return (float)(1.0-(t_d/i_d));
+    std::cerr<<(float)(1.0-(float)(t_d/i_d))<<std::endl;
+    return (1.0-(float)t_d/i_d);
 }
